revise __init__ usage
