+++ extras/devices/LVFS/tests/test_LVFS.py
@@ -23,7 +23,7 @@ class TestLVFSDevice(unittest.TestCase):
         Test if upgrade function returns True. And test if downgrade function
         returns False.
         """
-        with patch("devices.LVFS.LVFSDevice.run_cmd") as mock_path:
+        with patch("devices.LVFSDevice.run_cmd") as mock_path:
             mock_path.side_effect = self.mock_run_cmd
             device = LVFSDevice("", "", "")
             device._parse_fwupd_raw(fwupd_data.GET_DEVICES_RESPONSE_DATA)
@@ -36,7 +36,7 @@ class TestLVFSDevice(unittest.TestCase):
         Test if upgrade function returns False. And test if downgrade function
         returns True.
         """
-        with patch("devices.LVFS.LVFSDevice.run_cmd") as mock_path:
+        with patch("devices.LVFSDevice.run_cmd") as mock_path:
             mock_path.side_effect = self.mock_run_cmd
             device = LVFSDevice("", "", "")
 
@@ -53,7 +53,7 @@ class TestLVFSDevice(unittest.TestCase):
     def test_check_results_failed_state(self):
         """Validate UpdateState check in check_results"""
         global device_results
-        with patch("devices.LVFS.LVFSDevice.run_cmd") as mock_path:
+        with patch("devices.LVFSDevice.run_cmd") as mock_path:
             mock_path.side_effect = self.mock_run_cmd
             device = LVFSDevice("", "", "")
             device._parse_fwupd_raw(fwupd_data.GET_DEVICES_RESPONSE_DATA)
@@ -65,7 +65,7 @@ class TestLVFSDevice(unittest.TestCase):
     def test_check_results_mismatched_version(self):
         """Validate version check in check_results"""
         global device_results
-        with patch("devices.LVFS.LVFSDevice.run_cmd") as mock_path:
+        with patch("devices.LVFSDevice.run_cmd") as mock_path:
             mock_path.side_effect = self.mock_run_cmd
             device = LVFSDevice("", "", "")
             device._parse_fwupd_raw(fwupd_data.GET_DEVICES_RESPONSE_DATA)
@@ -75,7 +75,7 @@ class TestLVFSDevice(unittest.TestCase):
     def test_check_results_good(self):
         """Test if check_results works with a valid case"""
         global device_results
-        with patch("devices.LVFS.LVFSDevice.run_cmd") as mock_path:
+        with patch("devices.LVFSDevice.run_cmd") as mock_path:
             mock_path.side_effect = self.mock_run_cmd
             device = LVFSDevice("", "", "")
             device._parse_fwupd_raw(fwupd_data.GET_DEVICES_RESPONSE_DATA)
@@ -91,7 +91,7 @@ class TestLVFSDevice(unittest.TestCase):
         get-results```
         """
         global device_results
-        with patch("devices.LVFS.LVFSDevice.run_cmd") as mock_path:
+        with patch("devices.LVFSDevice.run_cmd") as mock_path:
             mock_path.side_effect = self.mock_run_cmd
             device = LVFSDevice("", "", "")
             device._parse_fwupd_raw(fwupd_data.GET_DEVICES_RESPONSE_DATA)
