+++ extras/devices/OEM/__init__.py
@@ -1,0 +1,6 @@
+from devices.OEM.OEM import OEMDevice, HPEDevice
+
+__all__ = [
+    "OEMDevice",
+    "HPEDevice",
+]
